requirements.txt
# 🚀 Core framework
fastapi==0.110.0
uvicorn[standard]==0.29.0

# 🤖 Machine Learning / NLP
transformers==4.39.3
accelerate==0.27.2
llama-cpp-python==0.2.11
huggingface_hub

# 📦 Data handling
numpy<2
python-multipart==0.0.9
pydantic==2.6.4
pymongo==4.6.1
python-dotenv==1.0.1
requests==2.32.3

# 🔑 Authentication
firebase-admin

# 🧪 Testing
pytest==8.0.0
